requirements.txt
# This file is autogenerated. Do not edit it manually.
retrying==1.3.3
requests==2.17.3
PyYAML==3.13
boto3==1.9.6

## The following requirements were added by pip freeze:
botocore==1.12.7
certifi==2018.8.24
chardet==3.0.4
docutils==0.14
idna==2.5
jmespath==0.9.3
python-dateutil==2.7.3
s3transfer==0.1.13
six==1.11.0
urllib3==1.21.1
